Started HUD UI + Safe Area Fix
Started work on implementing all the HUD UI and added a safe area fix so that the game screen will never be affected by phones with side obstructions like cameras and panel cutoffs. Now black bars appear on the sides of the screen so that nothing is ever cut off.