Link Apple ID + Edit Profile Name Implemented
Added the ability to link an Apple ID to a guest account from the profile screen (which can be accessed now by pressing your profile icon or name in the top left of the main menu). Also added the ability to change your profile name within the profile screen by pressing the edit button. Name is limited to 16 characters, and is given a # and 4 digit number suffix to distinguish player accounts.