Stage Select Update + In-Game Quit Works
Stage selection now works with player level and energy. A new stage is unlocked every 5 profile levels and the player needs 5 energy to play each level. The game now uses 5 energy when you play, and this lead me to fix the quit from in-game option, which would break when returning to the main menu. Now when you quit, the main menu skips the splash screen and takes you directly to stage select since you remain signed in. Your profile info, hat tiers, and upgrades are loaded, while your energy recharge timer updates based on the amount of time you were in the game.