Purchase Energy Screen
Created the UI for a purchase energy screen that will show when clicking on your energy in the top bar of the main menu. This will provide the player the option to spend coins, gems, or watch an ad to receive a set amount of energy. There will be a limit on the amount of times you can select each option a day.