Purchase Energy Functionality
Added functionality to the purchase energy UI. It now allows players to purchase 10 energy for 100 coins or 15 energy for 25 gems if they are lacking the energy. There is also an option for watching an ad for 5 gems. Players can use each option 3 times a day, and the game keeps track of this, resetting to 3 chances on each new day. Also added dev option to reset your chances for the day in the dev options panel.