Inventory Display Started
Created the framework for displaying your weapons, equipment, and hats in the inventory panel during the game. The player's current weapons and equipment just need to be fed into the UI manager for them to work. The hats need a better script that holds their sprite icon and tier before that can work.