Pitchfork Update
Added an animation through code for the pitchfork so that it pulls back, stabs, and retracts based on attack cooldown time. Also added a range variable so the distance the pitchfork extends can be increased when upgrading, and made this work with the animation. Pitchfork also only rotates when retracted now, and doesn't do the weird sprite flipping thing.