Victory Screen
Created the victory screen UI for beating a level, while also adding its functionality so that it can display your enemies defeated, coins collected, and profile XP earned. Animations work as well, so the only thing left to do is call a public function in the UI manager when the game is over.